chore: lint修正 - DisplayNameParser __all__追加、import整列、round不要int除去
parsers/__init__.pyのDisplayNameParser importをソート済み位置に移動し
__all__に追加。widgets.pyの冗長なint(round())をround()に簡素化。

+++ jj/services/dashboard/widgets.py
@@ -263,8 +263,8 @@ def apply_style_to_fig(fig: Any, style: dict[str, int]) -> None:
     if "font_size" in style:
         font_sz = style["font_size"]
         # title: 24/20 倍、legend: 16/20 倍
-        title_font_sz = int(round(font_sz * 24 / 20))
-        legend_font_sz = int(round(font_sz * 16 / 20))
+        title_font_sz = round(font_sz * 24 / 20)
+        legend_font_sz = round(font_sz * 16 / 20)
         fig.update_layout(
             font=dict(size=font_sz),
             title_font=dict(size=title_font_sz),
